modularisation

+++ raytracer/Source/Intersection.h
@@ -13,8 +13,8 @@ struct Intersection
     glm ::vec3 colour;
     float distance;
     vec4 normal;
-    float reflect;
-    float refract;
+    double reflect;
+    double refract;
 };
 
 vec3 solveLinearEq(Triangle triangle, vec4 start, vec4 dir);
